requirements.txt
requests[socks]
python-telegram-bot
flask
sqlalchemy
apscheduler
qrcode[pil]
python-dotenv